backend/requirements.txt
fastapi==0.112.0
uvicorn==0.30.6
websockets==13.0
anthropic==0.34.0
apscheduler==3.10.4
httpx==0.27.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
